+++ arrus/core/api/ops/us4r/constraints/TxRxSequenceLimits.h
@@ -12,24 +12,32 @@ class TxRxSequenceLimitsBuilder;
  */
 class TxRxSequenceLimits {
 public:
-    TxRxSequenceLimits(const TxRxLimits &txrx, const Interval<uint32> &size) : txrx(txrx), size(size) {}
+    TxRxSequenceLimits(const TxRxLimits &txrx, const Interval<uint32> &size, uint32 maxNFirings)
+        : txrx(txrx), size(size), maxNFirings(maxNFirings) {}
+
+    /** Assumes that the sequence size == max number of firings. */
+    TxRxSequenceLimits(const TxRxLimits &txrx, const Interval<uint32> &size): TxRxSequenceLimits(txrx, size, size.end()) {}
 
     const TxRxLimits &getTxRx() const { return txrx; }
     const Interval<uint32> &getSize() const { return size; }
+    uint32 getMaxNumberOfFirings() const { return maxNFirings; }
 
 private:
     friend class TxRxSequenceLimitsBuilder;
     TxRxLimits txrx;
     Interval<uint32> size;
+    /** Maximum number of different TX/RX operators.
+     * It should be equal to the min(max rx apertures, max tx apertures, max tx delays). */
+    uint32 maxNFirings;
 };
 
 class TxRxSequenceLimitsBuilder {
 public:
     explicit TxRxSequenceLimitsBuilder(TxRxSequenceLimits limits): limits(std::move(limits)) {}
 
-    TxRxSequenceLimitsBuilder& setTxRxLimits(const TxLimits &txLimits, const RxLimits &rxLimits,
+    TxRxSequenceLimitsBuilder& setTxRxLimits(const TxLimits &txLimits0, const TxLimits &txLimits1, const RxLimits &rxLimits,
                                           const Interval<float> &pri) {
-        limits->txrx = TxRxLimits{txLimits, rxLimits, pri};
+        limits->txrx = TxRxLimits{txLimits0, txLimits1, rxLimits, pri};
         return *this;
     }