+++ arrus/core/devices/us4r/Us4RFactoryImpl.h
@@ -72,6 +72,8 @@ public:
                 backplane.value()->enableInternalTrigger();
             }
 
+            auto txrxlimits = settings.getTxRxLimits();
+            
             auto [us4oems, masterIUs4OEM] =
                 getUs4OEMs(us4OEMSettings, settings.isExternalTrigger(), probeAdapterSettings.getIOSettings(),
                            ius4oemHandles, settings.getTxRxLimits());
@@ -91,14 +93,15 @@ public:
                 pValidator.throwOnErrors();
                 probeOrdinal++;
             }
+            bool isExternalTrigger = false;
             if(backplane.has_value() && settings.isExternalTrigger()) {
-                backplane.value()->enableExternalTrigger();
+                isExternalTrigger = true;
             }
             auto hv = getHV(settings.getHVSettings(), ius4oems, backplane);
             return std::make_unique<Us4RImpl>(
                 id, std::move(us4oems), std::move(probeSettings), std::move(adapterSettings), std::move(hv), rxSettings,
                 settings.getChannelsMaskForAllProbes(), std::move(backplane), settings.getBitstreams(),
-                isBitstreamAddr, adapterSettings.getIOSettings());
+                isBitstreamAddr, adapterSettings.getIOSettings(), isExternalTrigger) ;
         } else {
             throw IllegalArgumentException("Custom OEM configuration is not available since 0.11.0.");
         }