+++ arrus/core/devices/us4r/us4oem/Us4OEMDescriptorFactory.h
@@ -31,11 +31,17 @@ public:
                 0.5f,  // number of TX periods resolution
                 isMaster,
                 arrus::ops::us4r::TxRxSequenceLimits {
-                    arrus::ops::us4r::TxRxLimits {
+                    arrus::ops::us4r::TxRxLimits { // rail 0
                         arrus::ops::us4r::TxLimits {
                             Interval<float>{minFrequency, maxFrequency},  // Frequency
                             Interval<float>{0.0f, 16.96e-6f}, // delay
-                            Interval<float>{0.0f, (float)(32.0f/10e6)}, // pulse length,
+                            Interval<float>{0.5f, (float)(32.0f)}, // pulse length in cycles,
+                            Interval<Voltage>{5, 90}
+                        },
+                        arrus::ops::us4r::TxLimits { // rail 1
+                            Interval<float>{minFrequency, maxFrequency},  // Frequency
+                            Interval<float>{0.0f, 16.96e-6f}, // delay
+                            Interval<float>{0.5f, (float)(32.0f)}, // pulse length in cycles,
                             Interval<Voltage>{5, 90}
                         },
                         arrus::ops::us4r::RxLimits {
@@ -43,9 +49,11 @@ public:
                         },
                         Interval<float>{35e-6f, 1.0f},  // PRI, == (the sequence reprogramming time, 1s)
                     },
-                    Interval<uint32>{0, 16384} // sequence length
+                    Interval<uint32>{0, 16384}, // sequence length,
+                    1024 // maximum number of different TX/RXs
                 },
-                0 // maximum number of TX timeouts
+                0, // maximum number of TX timeouts
+                229 // sample number TX start (i.e. TX delay = 0)
             };
         case 2:
             // us4OEM+ variant 0
@@ -53,8 +61,8 @@ public:
                 version, // us4OEM version
                 32, // RX channels
                 20e-6f,  // min. RX time,
-                5e-6f, // RX time epsilon,
-                35e-6f, // TX parameters reprogramming time,
+                0e-6f, // RX time epsilon,
+                6e-6f, // TX parameters reprogramming time,
                 65e6f, // Sampling frequency [Hz]
                 1ull << 32u, // DDR memory size [B]
                 1ull << (14+12), // Max transfer size [B]
@@ -62,10 +70,16 @@ public:
                 isMaster,
                 arrus::ops::us4r::TxRxSequenceLimits {
                     arrus::ops::us4r::TxRxLimits {
-                        arrus::ops::us4r::TxLimits {
+                        arrus::ops::us4r::TxLimits { // rail 0
                             Interval<float>{minFrequency, maxFrequency},  // Frequency
                             Interval<float>{0.0f, 16.96e-6f}, // delay
-                            Interval<float>{0.0f, (float)(32.0f/10e6f)}, // pulse length,
+                            Interval<float>{0.0f, (float)(32.0f)}, // pulse length in cycles,
+                            Interval<Voltage>{5, 90}
+                        },
+                        arrus::ops::us4r::TxLimits { // rail 1
+                            Interval<float>{minFrequency, maxFrequency},  // Frequency
+                            Interval<float>{0.0f, 16.96e-6f}, // delay
+                            Interval<float>{0.0f, (float)(32.0f)}, // pulse length in cycles,
                             Interval<Voltage>{5, 90}
                         },
                         arrus::ops::us4r::RxLimits {
@@ -73,9 +87,11 @@ public:
                         },
                         Interval<float>{35e-6f, 1.0f},  // PRI, == (the sequence reprogramming time, 1s)
                     },
-                    Interval<uint32>{0, 4096} // sequence length
+                    Interval<uint32>{0, 4096}, // sequence length
+                    4096 // maximum number of different TX/RXs
                 },
-                4 // maximum number of timeouts
+                4, // maximum number of timeouts
+                35 // sample number TX start (i.e. TX delay = 0)
             };
         default:
             throw arrus::IllegalArgumentException(format("Unsupported us4OEM version: {}", version));