+++ arrus/core/devices/us4r/us4oem/Us4OEMImpl.h
@@ -81,11 +81,12 @@ public:
     float getUCDTemperature() override;
     float getUCDExternalTemperature() override;
     float getUCDMeasuredVoltage(uint8_t rail) override;
+    float getMeasuredHVPVoltage() override;
+    float getMeasuredHVMVoltage() override;
     void checkFirmwareVersion() override;
     uint32 getFirmwareVersion() override;
     void checkState() override;
     uint32 getTxFirmwareVersion() override;
-    uint32_t getTxOffset() override;
     uint32_t getOemVersion() override;
     void setTestPattern(RxTestPattern pattern) override;
     uint16_t getAfe(uint8_t address) override;